docs(site): generalize untrusted-history guidance beyond snapshots
Move the guidance into a dedicated Trusted Message History page covering
all three ways history enters an agent (direct invocation, session restore,
snapshot load), and link to it with short callouts from the Prompts,
Session Management, and Snapshots pages.
