+++ site/src/content/docs/user-guide/concepts/agents/snapshots_imports.ts
@@ -7,8 +7,3 @@ import { Agent } from '@strands-agents/sdk'
 // --8<-- [start:imports_snapshot]
 import { Agent, type Snapshot } from '@strands-agents/sdk'
 // --8<-- [end:imports_snapshot]
-
-// --8<-- [start:clear_trailing_tool_use_imports]
-import { Agent } from '@strands-agents/sdk'
-import type { MessageData, Snapshot } from '@strands-agents/sdk'
-// --8<-- [end:clear_trailing_tool_use_imports]
