+++ site/src/content/docs/user-guide/safety-security/trusted-message-history_imports.ts
@@ -1,0 +1,4 @@
+// --8<-- [start:clear_trailing_tool_use_imports]
+import { Agent } from '@strands-agents/sdk'
+import type { MessageData } from '@strands-agents/sdk'
+// --8<-- [end:clear_trailing_tool_use_imports]
